LayoutTests/fast/dom/HTMLDialogElement/non-modal-dialog-does-not-block-mouse-events-expected.txt

Test for bug 110952. Ensure that non-modal dialogs do not block mouse events. To test manually, click the red box. The test succeeds if the red box turns green.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS div.firedOn is true
PASS successfullyParsed is true

TEST COMPLETE
